Allow nicer API for loading multiple fixtures of the same type, bump version to 1.2.0
This commit allows you to change the following code:

  FixtureDependencies.load(:model__obj1, :model__obj2, :model__obj3)

to:

  FixtureDependencies.load(:models=>[:obj1, :obj2, :obj3])

Basically, it makes the API for loading multiple fixtures of the same
type much easier, without having to load all fixtures of that type
at once (which makes for brittle tests).

Additionally, it adds a helper for RSpec and Sequel, giving easy
access to the load method and running each spec in separate
transaction.