Add composite key support when using Sequel
This is based on a request and initial work from Greg Ditrick.
It modifies the code to check for composite keys in a few places,
and to handle them appropriately.  Composite key support works
almost exactly like regular key support.  In both cases, you
should use the symbolic names to reference the associated objects.