Add `use_unsafe_load` setting
Since Ruby 3.1, Psyche (the gem in charge of parsing YAMLs) tries
parses certain string patterns as `Date` or `Time` objects, but fails to
instantiate them if those classes are not explicitly allowed, or you use
`unsafe_load` instead of `load`.

With this change users of `fixture_dependencies` can opt in to the
`unsafe_load` way of solving this issue.

Closes: #37
Closes: #38

Co-Authored-By: Jeremy Evans <code@jeremyevans.net>