Make FixtureDependencies.class_map handle model names that are not CamelCase (Fixes #39)
For most of this, just pass the mapped class along as an option, which
isn't too bad.

However, for associations, we need an inverse mapping of class to symbol.
As the class_map can be modified at any time, this currently uses Hash#find,
which is suboptimal for performance.  However, as long as class_map is
small, it shouldn't cause a significant performance issue.  If it is a
performance issue, we can support frozen class maps, and then we can build
an inverse_class_map and index into that.